refactor(hygiene): probe 护栏收敛为 hygiene_probe example + 新增 COMPONENT_GUIDE
external-api-probe 转成 examples/hygiene_probe.rs(ratatui-kit-examples 不直连 ratatui/crossterm,一个 example 即可当 hygiene 护栏),移除两个顶层 probe crate;rename 逃生舱验证降级为文档(extern crate 是 Rust 语言机制)。新增 COMPONENT_GUIDE.md 第三方作者指南,更新 EXTENSION_API.md 与知识库的护栏引用。
